pyproject.toml
[project]
name = "a2a-examples"
version = "0.1.0"
description = "Agent-to-Agent protocol examples"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "a2a-sdk[sql]>=0.3.0",
    "crewai[tools]>=0.95.0",
    "fastmcp>=1.0",
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.91.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.5",
    "openai>=2.2.0",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.1",
    "litellm",
    "matplotlib>=3.8.0",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
]
